Backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.5
django-filter==24.3
django-jazzmin==3.0.1
django-rest-passwordreset==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
